refact: rename commands/controllers for consistency
