+++ lib/controllers/commands/commands.dart
@@ -1,6 +1,6 @@
 export 'launch_url_command.dart';
-export 'save_questionnaire_command.dart';
-export 'submit_questionnaire_command.dart';
+export 'save_submit/save_responses_command.dart';
+export 'save_submit/submit_questionnaire_command.dart';
 export 'update_response/debounce_and_save_response_command.dart';
 export 'update_response/toggle_checkbox_command.dart';
 export 'update_response/toggle_decline_to_respond_command.dart';
