+++ lib/controllers/commands/save_submit/save_responses_command.dart
@@ -1,8 +1,8 @@
 import 'package:prapare/models/fhir_questionnaire/survey/export.dart';
 
-import 'abstract_command.dart';
+import '../abstract_command.dart';
 
-class SubmitQuestionnaireCommand extends AbstractCommand {
+class SaveResponsesCommand extends AbstractCommand {
   @override
   Future<void> execute() async {
     final answerListObs = responsesController.userResponsesMap.values;
